# backend/src/public/swagger.json
{
	"components": {
...
		"schemas": {
			"Ingredient": {
				"properties": {
					"name": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"taxedValue": {
						"type": "number",
						"format": "double"
					},
					"longName": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"value",
					"taxedValue",
					"longName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IngredientDrop": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					},
					"ingredient": {
						"$ref": "#/components/schemas/Ingredient"
					}
				},
				"required": [
					"amount",
					"ingredient"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Berry": {
				"properties": {
					"name": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MainSkill": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pokemon": {
				"properties": {
					"name": {
						"type": "string"
					},
					"specialty": {
						"type": "string",
						"enum": [
							"berry",
							"ingredient",
							"skill"
						]
					},
					"frequency": {
						"type": "number",
						"format": "double"
					},
					"ingredientPercentage": {
						"type": "number",
						"format": "double"
					},
					"skillPercentage": {
						"type": "number",
						"format": "double"
					},
					"berry": {
						"$ref": "#/components/schemas/Berry"
					},
					"carrySize": {
						"type": "number",
						"format": "double"
					},
					"maxCarrySize": {
						"type": "number",
						"format": "double"
					},
					"ingredient0": {
						"$ref": "#/components/schemas/IngredientDrop"
					},
					"ingredient30": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					},
					"ingredient60": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					},
					"skill": {
						"$ref": "#/components/schemas/MainSkill"
					}
				},
				"required": [
					"name",
					"specialty",
					"frequency",
					"ingredientPercentage",
					"skillPercentage",
					"berry",
					"carrySize",
					"maxCarrySize",
					"ingredient0",
					"ingredient30",
					"ingredient60",
					"skill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PokemonCombination": {
				"properties": {
					"ingredientList": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					},
					"pokemon": {
						"$ref": "#/components/schemas/Pokemon"
					}
				},
				"required": [
					"ingredientList",
					"pokemon"
				],
				"type": "object"
			},
...
			"Meal": {
				"properties": {
					"name": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"value50": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string",
						"enum": [
							"curry",
							"salad",
							"dessert"
						]
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					},
					"bonus": {
						"type": "number",
						"format": "double"
					},
					"unlockables": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"value",
					"value50",
					"type",
					"ingredients",
					"bonus",
					"unlockables"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Contribution": {
				"properties": {
					"meal": {
						"$ref": "#/components/schemas/Meal"
					},
					"percentage": {
						"type": "number",
						"format": "double"
					},
					"contributedPower": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"meal",
					"percentage",
					"contributedPower"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CombinedContribution": {
				"properties": {
					"bestMeals": {
						"items": {
							"$ref": "#/components/schemas/Contribution"
						},
						"type": "array"
					},
					"averagePercentage": {
						"type": "number",
						"format": "double"
					},
					"summedContributedPower": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"bestMeals",
					"averagePercentage",
					"summedContributedPower"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PokemonCombinationContribution": {
				"properties": {
					"pokemonCombination": {
						"$ref": "#/components/schemas/PokemonCombination"
					},
					"combinedContribution": {
						"$ref": "#/components/schemas/CombinedContribution"
					}
				},
				"required": [
					"pokemonCombination",
					"combinedContribution"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TierlistType": {
				"type": "string",
				"enum": [
					"overall",
					"curry",
					"salad",
					"dessert"
				]
			},
...
			"CombinationForFlexibleRankingType": {
				"properties": {
					"pokemon": {
						"type": "string"
					},
					"averagePercentage": {
						"type": "number",
						"format": "double"
					},
					"ingredientList": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					}
				},
				"required": [
					"pokemon",
					"averagePercentage",
					"ingredientList"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CombinationForFocusedRankingType": {
				"properties": {
					"pokemon": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"meals": {
						"type": "string"
					},
					"ingredientList": {
						"items": {
							"$ref": "#/components/schemas/IngredientDrop"
						},
						"type": "array"
					}
				},
				"required": [
					"pokemon",
					"total",
					"meals",
					"ingredientList"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sleep-api",
		"version": "5.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/tierlist/cooking": {
			"get": {
				"operationId": "GetCookingTierlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PokemonCombinationContribution"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tierlistType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/TierlistType"
						}
					},
					{
						"in": "query",
						"name": "limit50",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "pretty",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "onlyBest",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "previous",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/ranking/meal/flexible": {
			"get": {
				"operationId": "GetMealGeneralistRankingRaw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CombinationForFlexibleRankingType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit30",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "cyan",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "taupe",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "snowdrop",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "lapis",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "advanced",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "unlocked",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "lategame",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "curry",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "salad",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "dessert",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "pretty",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "csv",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/ranking/meal/focused": {
			"get": {
				"operationId": "GetMealFocusedRankingRaw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CombinationForFocusedRankingType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit30",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "cyan",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "taupe",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "snowdrop",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "lapis",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "advanced",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "unlocked",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "lategame",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "curry",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "salad",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "dessert",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "pretty",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "csv",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "nrOfMeals",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}